revisi 1

+++ application/views/administrator/mod_kegiatan/Data.php
@@ -29,6 +29,7 @@
     								<th>NO</th>
     								<th>Judul Kegiatan</th>
                                     <th>Jenis Kegiatan</th>
+                                    <th>Penulis</th>
     								<th>Actions</th>
     							</tr>
     						</thead>
@@ -40,14 +41,19 @@
     									<td><?=$no?></td>
     									<td><?=$l['judul_kegiatan']?></td>
                                         <td><?=$l['jenis_kegiatan']?></td>
+                                        <td><?=$l['nama']?></td>
     									<td>
-    										<div class="btn-group">
-    											<a href="<?=base_url()?>administrator/edit_kegiatan/<?=$l['id_kegiatan']?>?lembaga=<?=$this->session->userdata('username')?>" class="btn btn-success">Edit</a>
-    											
-    											<button type="button" class="btn btn-default" disabled=""><i class="fa fa-align-center"></i></button>
+                                            <?php if ($l['author'] == $this->session->userdata('user')) { ?>
+                                                <div class="btn-group">
+                                                    <a href="<?=base_url()?>administrator/edit_kegiatan/<?=$l['id_kegiatan']?>?lembaga=<?=$this->session->userdata('username')?>" class="btn btn-success">Edit</a>
 
-    											<a href="<?=base_url()?>administrator/hapus_kegiatan/<?=$l['id_kegiatan']?>?lembaga=<?=$this->session->userdata('username')?>" class="btn btn-danger">Hapus</a>
-    										</div>
+                                                    <button type="button" class="btn btn-default" disabled=""><i class="fa fa-align-center"></i></button>
+
+                                                    <a href="<?=base_url()?>administrator/hapus_kegiatan/<?=$l['id_kegiatan']?>?lembaga=<?=$this->session->userdata('username')?>" class="btn btn-danger">Hapus</a>
+                                                </div>
+                                            <?php } else { ?>
+                                                <p>Anda Tidak Memiliki Akses</p>
+                                            <?php } ?>    										
     									</td>
     								</tr>
     							<?php $no++; } ?>
